+++ src/zepben/eas/client/eas_client.py
@@ -244,7 +244,12 @@ class EasClient:
                 "useSpanLevelThreshold": generator_config.model.use_span_level_threshold,
                 "ratingThreshold": generator_config.model.rating_threshold,
                 "simplifyPLSIThreshold": generator_config.model.simplify_plsi_threshold,
-                "emergAmpScaling": generator_config.model.emerg_amp_scaling
+                "emergAmpScaling": generator_config.model.emerg_amp_scaling,
+                "inverterControlConfig": generator_config.model.inverter_control_config and {
+                    "cutOffDate": generator_config.model.inverter_control_config.cut_off_date and generator_config.model.inverter_control_config.cut_off_date.isoformat(),
+                    "beforeCutOffProfile": generator_config.model.inverter_control_config.beforeCutOffProfile,
+                    "afterCutOffProfile": generator_config.model.inverter_control_config.afterCutOffProfile
+                }
             },
             "solve": generator_config.solve and {
                 "normVMinPu": generator_config.solve.norm_vmin_pu,
@@ -1330,6 +1335,11 @@ class EasClient:
                                             ratingThreshold
                                             simplifyPLSIThreshold
                                             emergAmpScaling
+                                            inverterControlConfig {
+                                                cutOffDate
+                                                beforeCutOffProfile
+                                                afterCutOffProfile
+                                            }
                                         }
                                         solve {
                                             normVMinPu